refactor: HttpMessage -> HttpRequest로 이름 변경
